Modify Commands and Ui to fix find and delete feature
