+++ src/main/java/dukeclasses/Commands.java
@@ -43,7 +43,7 @@ public enum Commands {
             return "    Noted. I've removed this task:\n";
         }
     },
-    FIND{
+    MATCH{
         @Override
         public String toString(){
             return "    Here are the matching tasks in your list:";
