+++ src/main/java/dukeclasses/Ui.java
@@ -5,14 +5,15 @@ import java.io.FileNotFoundException;
 import java.util.Scanner;
 
 /**
- * Represents a Ui which deal with returning results to users through the use of print statements.
+ * Represents a Ui
  */
 public class Ui {
 
     /**
-     * Class constructor to instantiate Ui.
+     * Class constructor to initiate Ui
      */
     public Ui(){
+
     }
 
     /**
@@ -23,71 +24,77 @@ public class Ui {
     }
 
     /**
-     * Prints error if file does not exist when trying to read or write.
+     * Prints error if file is invalid.
      */
     public void showFileError() {
         System.out.println("Invalid File");
     }
 
     /**
-     * Prints error if storage fail to update when list of tasks are modified.
+     * Prints error if storage fail to updates.
      */
     public void showStorageError() {
         System.out.println("Fail to update storage");
     }
 
     /**
-     * Prints error if user input does not follow format.
+     * Prints error if user gives an invalid input.
      */
     public void showInputError() {
         System.out.println("    Invalid input detected. Please check your input");
     }
 
     /**
-     * Prints greeting during boot up or when hi command is called.
+     * Prints greeting message.Usually immediately after boot up or when called using 'hi' command.
      */
     public void greet() {
         System.out.println(Commands.HI.toString());
     }
 
     /**
-     * Prints message when user initiate exit command.
+     * Prints bye message before the program closes.
      */
     public void sayBye() {
         System.out.println(Commands.BYE.toString());
     }
 
     /**
-     * Prints details of the task.
+     * Prints a message indicating that task mentioned is not found.
+     */
+    public void sayTaskNotFound() {
+        System.out.println("Task not found.");
+    }
+
+    /**
+     * Prints the identity of the given task.
      *
      * @param task Task which will have its details printed.
      */
     public void identifyTask(Task task) {
         if (task.getIsDone()) {
-            String output = String.format(
-                "%s    %s", Commands.MARK.toString(), task.identify());
+            String output = String.format("%s    %s", Commands.MARK.toString(), task.identify());
             System.out.println(output);
         } else {
-            System.out.println(String.format(
-                "%s      %s", Commands.UNMARK.toString(), task.identify()));
+            System.out.println(String.format("%s      %s", Commands.UNMARK.toString(), task.identify()));
         }
     }
 
     /**
-     * Prints details of the newly added task and also current number of tasks in list.
+     * Prints message indicating new task was added.
      *
-     * @param task Task which will have its details printed.
-     * @param numOfTask Integer that denotes number of tasks in the list currently.
+     * @param task Task that was added.
+     * @param numOfTask Number of tasks in the list.
      */
     public void newTask(Task task, int numOfTask) {
-        System.out.println(String.format("%s      %s", Commands.ADD.toString(), task.identify()));
+        System.out.println(String.format(
+            "%s      %s", Commands.ADD.toString(), task.identify()));
         System.out.println(String.format("    Now you have %d tasks in the list.", numOfTask));
     }
 
     /**
-     * Prints list of task.
+     * Prints list of task in storage file.
      *
-     * @param filePath File which contain list of task.
+     * @param filePath Path of storage file.
      */
     public void listTask(String filePath) {
         System.out.println(Commands.LIST.toString());
@@ -104,19 +111,25 @@ public class Ui {
     }
 
     /**
-     * Prints details of the deleted task.
+     * Prints list of task with the given tasklist.
      *
-     * @param task Task which will have its details printed.
+     * @param tasks Tasklist which will have its tasks printed.
      */
-    public void deleteTask(Task task) {
-        System.out.println(String.format("%s      %s", Commands.DELETE.toString(), task.identify()));
-    }
-
     public void listTaskUsingArrayList(TaskList tasks) {
-        System.out.println(Commands.FIND.toString());
+        System.out.println(Commands.MATCH.toString());
         for (int i = 0; i < tasks.getTaskList().size(); i++) {
             Task task = tasks.getTaskList().get(i);
             System.out.println(String.format("    %s", task.identify()));
         }
+    }
+
+    /**
+     * Prints message when task is deleted.
+     *
+     * @param task Task that was deleted.
+     */
+    public void deleteTask(Task task) {
+        System.out.println(String.format(
+                "%s      %s", Commands.DELETE.toString(), task.identify()));
     }
 }
